pyproject.toml
[tool.poetry]
name = "ra_graphql"
version = "0.1.0"
description = ""
authors = ["Shubham Biswas <shubhambiswas.zero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.59.0"
uvicorn = "^0.11.5"
ariadne = "^0.11.0"
alembic = "^1.4.2"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8.5"
pytz = "^2020.1"
passlib = "^1.7.2"
bcrypt = "^3.1.7"
pyjwt = "^1.7.1"
email_validator = "^1.1.1"
permissions_system = "^0.2"
autopep8 = "^1.5.3"
databases = {extras = ["postgresql"], version = "^0.3.2"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
